Add post request demo for API
